refactor: 기존 페이지 컴포넌트를 이름을 변경
뷰 라우터 추가에 따라 기존 Page 이름을 가진 컴포넌트를 Section이라는 이름으로 변경함.
