+++ frontend/src/App.vue
@@ -1,23 +1,23 @@
 <script>
-import ProductPage from "@/component/product/ProductPage.vue";
-import OrderPage from "@/component/order/OrderPage.vue";
+import ProductSector from "@/component/product/ProductSector.vue";
+import OrderSector from "@/component/order/OrderSector.vue";
 import AddProductDialog from "@/component/product/AddProductDialog.vue";
 
 const ORDER = 'ORDER';
 const PRODUCT = 'PRODUCT';
 export default {
   name: "App",
-  components: {AddProductDialog, OrderPage, ProductPage},
+  components: {AddProductDialog, OrderSector, ProductSector},
   data() {
     return {
       page: ORDER,
     }
   },
   methods: {
-    selectOrderPage() {
+    selectOrderSector() {
       this.page = ORDER;
     },
-    selectProductPage() {
+    selectProductSector() {
       this.page = PRODUCT;
     }
   }
@@ -29,8 +29,8 @@ export default {
     <v-container>
       <v-row no-gutters class="align-center">
         <v-col cols="3">
-          <div class="link" @click="selectOrderPage" v-if="page === 'PRODUCT'">주문 관리</div>
-          <div class="link" @click="selectProductPage" v-if="page === 'ORDER'">상품 관리</div>
+          <div class="link" @click="selectOrderSector" v-if="page === 'PRODUCT'">주문 관리</div>
+          <div class="link" @click="selectProductSector" v-if="page === 'ORDER'">상품 관리</div>
         </v-col>
         <v-col cols="6">
           <div class="text-h6 text-center" v-if="page === 'ORDER'">정글 주문 관리</div>
@@ -41,8 +41,8 @@ export default {
         </v-col>
       </v-row>
     </v-container>
-    <ProductPage v-if="page === 'PRODUCT'"/>
-    <OrderPage v-if="page === 'ORDER'"/>
+    <ProductSector v-if="page === 'PRODUCT'"/>
+    <OrderSector v-if="page === 'ORDER'"/>
   </div>
 </template>
