+++ frontend/src/component/product/ProductSector.vue
